generators/requirements.txt
kafka-python==2.0.2
faker==25.9.1
click==8.1.7
